+++ include/franka/robot.h
@@ -130,8 +130,8 @@ class Robot {
    *
    * @param[in] control_callback Callback function providing joint-level torque commands.
    * See @ref callback-docs "here" for more details.
-   * @param[in] saturate Flag for activating saturation of control commands. True by default.
-   * Saturation could distort your motion!
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to torque control or motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -141,7 +141,7 @@ class Robot {
    * @see Robot::Robot to change behavior if realtime priority cannot be set.
    */
   void control(std::function<Torques(const RobotState&, franka::Duration)> control_callback,
-               const bool saturate = true);
+               const bool limit_rate = true);
 
   /**
    * Starts a control loop for sending joint-level torque commands and joint positions.
@@ -153,8 +153,8 @@ class Robot {
    * See @ref callback-docs "here" for more details.
    * @param[in] motion_generator_callback Callback function for motion generation. See @ref
    * callback-docs "here" for more details.
-   * @param[in] saturate Flag for activating saturation of motion generator and control commands.
-   * True by default. Saturation could distort your motion!
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to torque control or motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -166,7 +166,7 @@ class Robot {
   void control(
       std::function<Torques(const RobotState&, franka::Duration)> control_callback,
       std::function<JointPositions(const RobotState&, franka::Duration)> motion_generator_callback,
-      const bool saturate = true);
+      const bool limit_rate = true);
 
   /**
    * Starts a control loop for sending joint-level torque commands and joint velocities.
@@ -178,8 +178,8 @@ class Robot {
    * See @ref callback-docs "here" for more details.
    * @param[in] motion_generator_callback Callback function for motion generation. See @ref
    * callback-docs "here" for more details.
-   * @param[in] saturate Flag for activating saturation of motion generator and control commands.
-   * True by default. Saturation could distort your motion!
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to torque control or motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -191,7 +191,7 @@ class Robot {
   void control(
       std::function<Torques(const RobotState&, franka::Duration)> control_callback,
       std::function<JointVelocities(const RobotState&, franka::Duration)> motion_generator_callback,
-      const bool saturate = true);
+      const bool limit_rate = true);
 
   /**
    * Starts a control loop for sending joint-level torque commands and Cartesian poses.
@@ -203,8 +203,8 @@ class Robot {
    * See @ref callback-docs "here" for more details.
    * @param[in] motion_generator_callback Callback function for motion generation. See @ref
    * callback-docs "here" for more details.
-   * @param[in] saturate Flag for activating saturation of motion generator and control commands.
-   * True by default. Saturation could distort your motion!
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to torque control or motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -216,7 +216,7 @@ class Robot {
   void control(
       std::function<Torques(const RobotState&, franka::Duration)> control_callback,
       std::function<CartesianPose(const RobotState&, franka::Duration)> motion_generator_callback,
-      const bool saturate = true);
+      const bool limit_rate = true);
 
   /**
    * Starts a control loop for sending joint-level torque commands and Cartesian velocities.
@@ -228,8 +228,8 @@ class Robot {
    * See @ref callback-docs "here" for more details.
    * @param[in] motion_generator_callback Callback function for motion generation. See @ref
    * callback-docs "here" for more details.
-   * @param[in] saturate Flag for activating saturation of motion generator and control commands.
-   * True by default. Saturation could distort your motion!
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to torque control or motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -241,7 +241,7 @@ class Robot {
   void control(std::function<Torques(const RobotState&, franka::Duration)> control_callback,
                std::function<CartesianVelocities(const RobotState&, franka::Duration)>
                    motion_generator_callback,
-               const bool saturate = true);
+               const bool limit_rate = true);
 
   /**
    * Starts a control loop for a joint position motion generator with a given controller mode.
@@ -252,8 +252,8 @@ class Robot {
    * @param[in] motion_generator_callback Callback function for motion generation. See @ref
    * callback-docs "here" for more details.
    * @param[in] controller_mode Controller to use to execute the motion.
-   * @param[in] saturate Flag for activating saturation of motion generator commands. True by
-   * default.
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -265,7 +265,7 @@ class Robot {
   void control(
       std::function<JointPositions(const RobotState&, franka::Duration)> motion_generator_callback,
       ControllerMode controller_mode = ControllerMode::kJointImpedance,
-      const bool saturate = true);
+      const bool limit_rate = true);
 
   /**
    * Starts a control loop for a joint velocity motion generator with a given controller mode.
@@ -276,8 +276,8 @@ class Robot {
    * @param[in] motion_generator_callback Callback function for motion generation. See @ref
    * callback-docs "here" for more details.
    * @param[in] controller_mode Controller to use to execute the motion.
-   * @param[in] saturate Flag for activating saturation of motion generator commands. True by
-   * default.
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -289,7 +289,7 @@ class Robot {
   void control(
       std::function<JointVelocities(const RobotState&, franka::Duration)> motion_generator_callback,
       ControllerMode controller_mode = ControllerMode::kJointImpedance,
-      const bool saturate = true);
+      const bool limit_rate = true);
 
   /**
    * Starts a control loop for a Cartesian pose motion generator with a given controller mode.
@@ -300,8 +300,8 @@ class Robot {
    * @param[in] motion_generator_callback Callback function for motion generation. See @ref
    * callback-docs "here" for more details.
    * @param[in] controller_mode Controller to use to execute the motion.
-   * @param[in] saturate Flag for activating saturation of motion generator commands. True by
-   * default.
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -313,7 +313,7 @@ class Robot {
   void control(
       std::function<CartesianPose(const RobotState&, franka::Duration)> motion_generator_callback,
       ControllerMode controller_mode = ControllerMode::kJointImpedance,
-      const bool saturate = true);
+      const bool limit_rate = true);
 
   /**
    * Starts a control loop for a Cartesian velocity motion generator with a given controller mode.
@@ -324,8 +324,8 @@ class Robot {
    * @param[in] motion_generator_callback Callback function for motion generation. See @ref
    * callback-docs "here" for more details.
    * @param[in] controller_mode Controller to use to execute the motion.
-   * @param[in] saturate Flag for activating saturation of motion generator commands. True by
-   * default.
+   * @param[in] limit_rate True if rate limiting should be activated. True by default.
+   * This could distort your motion!
    *
    * @throw ControlException if an error related to motion generation occurred.
    * @throw InvalidOperationException if a conflicting operation is already running.
@@ -337,7 +337,7 @@ class Robot {
   void control(std::function<CartesianVelocities(const RobotState&, franka::Duration)>
                    motion_generator_callback,
                ControllerMode controller_mode = ControllerMode::kJointImpedance,
-               const bool saturate = true);
+               const bool limit_rate = true);
 
   /**
    * @}